widget-editor: better separation of utils and constants
