+++ constants/widget-editor.js
@@ -1,10 +1,8 @@
-import { AdapterModifier } from '@widget-editor/widget-editor';
-import RWAdapter from '@widget-editor/rw-adapter';
-
 export const WIDGET_EDITOR_DEFAULT_DISABLED_FEATURES = [
   'typography',
   'end-user-filters',
 ];
+
 export const WIDGET_EDITOR_COLOUR_SCHEMES = [
   {
     name: 'Resource Watch',
@@ -50,9 +48,37 @@ export const WIDGET_EDITOR_COLOUR_SCHEMES = [
   },
 ];
 
-export const getRWAdapter = (config = {}) => AdapterModifier(RWAdapter, {
-  endpoint: process.env.WRI_API_URL,
-  env: process.env.API_ENV,
-  applications: process.env.APPLICATIONS.split(','),
-  ...config,
-});
+export const WIDGET_EDITOR_DEFAULT_THEME = {
+  height: 0, // Don't touch this without testing all the charts
+  // and particularly the bar chart with or without
+  // scrolling and its vertical alignment
+  padding: 'auto', // Do not set something different than 'auto'
+  // because it will break several graphs
+  // (primarily the bar and pie ones)
+  render: {
+    retina: true,
+  },
+  marks: {
+    color: '#1f77b4',
+  },
+  axis_x: {
+    axisColor: '#A9ABAD',
+    tickSize: 8,
+    ticks: 8,
+    tickColor: '#A9ABAD',
+    tickLabelColor: '#717171',
+    tickLabelFont: 'Lato',
+    tickLabelFontSize: 13,
+  },
+  axis_y: {
+    axisWidth: 0,
+    tickSize: 0,
+    ticks: 5,
+    tickLabelColor: '#717171',
+    tickLabelFont: 'Lato',
+    tickLabelFontSize: 13,
+    grid: true,
+    gridColor: '#A9ABAD',
+    gridOpacity: 1,
+  },
+};
